modernizing code
