+++ TODO.md
@@ -1,0 +1,26 @@
+# TODO
+
+## For 0.1 release
+
+* why does `print(hi)` return true when `hi` is undefined?
+* fix the segfault issues (superrecursion and ^C-ing)
+* clean up code, starting with parser.cpp
+  * continue where left off, line ~= 160
+* get rid of duplicate code in eval.cpp
+  * merge the functions into one
+* improve math funcs
+  * add more, like rng
+  * also make logical and bitwise operators new code
+* string manipulation
+  * access
+  * mutate
+  * concat
+  * split
+  * replace
+* implement `exit()` function
+
+## For 0.2 release
+
+* hex and binary support
+* arrays
+* hashmaps / objects
