+++ src/lexer.h
@@ -6,13 +6,12 @@
 
 #include "ShimmerClasses.h"
 
-typedef enum {UNKNOWN = -1, NONE, COMMENT, ID, INT, STR} LexerState;
+typedef enum {UNKNOWN = -1, NONE, COMMENT, ID, NUMBER, STR} LexerState;
 std::string str_repr(LexerState ls);
 
 typedef char chtype;
 
 std::vector<ShimmerToken> lex(std::string str);
-chtype esc_seq(chtype ch);
 
 ShimmerToken make_token(LexerState now_in, std::string current_token_contents, int line);
